PFA-28 Move typescript config back into separate backend folder
