+++ backend/tsconfig.json
@@ -4,7 +4,7 @@
     "module": "commonjs",
     "lib": ["ES2020"],
     "outDir": "./dist",
-    "rootDir": "../",
+    "rootDir": "./",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
@@ -15,14 +15,11 @@
     "sourceMap": true
   },
   "include": [
-    "./**/*.ts",
-    "../shared/**/*.ts"
+    "**/*.ts"
   ],
   "exclude": [
     "node_modules",
     "dist",
-    "test/**/*.js",
-    "../frontend/**/*",
-    "../node_modules"
+    "test/**/*.js"
   ]
 }
