+++ shared/types/expense.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=expense.js.map
