--- shared/types/expense.ts
@@ -1,38 +1,0 @@
-// Shared expense constants and types
-
-export const EXPENSE_CATEGORIES = {
-    HOUSING: 'Housing',
-    TRANSPORTATION: 'Transportation',
-    FOOD: 'Food',
-    HEALTHCARE: 'Healthcare',
-    ENTERTAINMENT: 'Entertainment',
-    SHOPPING: 'Shopping',
-    BILLS: 'Bills',
-    EDUCATION: 'Education',
-    TRAVEL: 'Travel',
-    OTHER: 'Other'
-} as const;
-
-export const RECURRING_FREQUENCIES = {
-    WEEKLY: 'Weekly',
-    BI_WEEKLY: 'Bi-weekly',
-    MONTHLY: 'Monthly',
-    QUARTERLY: 'Quarterly',
-    YEARLY: 'Yearly'
-} as const;
-
-// Utility types
-export type ExpenseCategory = typeof EXPENSE_CATEGORIES[keyof typeof EXPENSE_CATEGORIES];
-export type RecurringFrequency = typeof RECURRING_FREQUENCIES[keyof typeof RECURRING_FREQUENCIES];
-
-// Base expense interface (shared between frontend and backend)
-export interface BaseExpense {
-    amount: number;
-    dateSpent: Date | string; // Allow both Date and string for API compatibility
-    description?: string;
-    category: ExpenseCategory;
-    merchant?: string;
-    isRecurring: boolean;
-    recurringFrequency?: RecurringFrequency;
-    startDate?: Date | string; // Allow both Date and string for API compatibility
-}
